% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bland_mean.R
\name{bland_mean}
\alias{bland_mean}
\title{Estimate Mean from Quartiles and Range}
\usage{
bland_mean(a, q_1, m, q_3, b)
}
\arguments{
\item{a}{Minimum value of sample.}

\item{q_1}{First quartile of sample.}

\item{m}{Median of sample.}

\item{q_3}{Third quartile of sample.}

\item{b}{Maximum value of sample.}
}
\description{
See \code{\link{estimator_eqn_ref}} for details and references for equation
(10).
}
