% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_cauchy.R
\name{g_cauchy}
\alias{g_cauchy}
\title{Estimator assuming a Cauchy distribution.}
\usage{
g_cauchy(n, median, spread, spread_type = "iqr")
}
\arguments{
\item{n}{sample size}

\item{median}{sample median}

\item{spread}{iqr or range value}

\item{spread_type}{iqr or range, defaults to iqr}
}
\description{
This estimator is for comparison purposes in the computational exploration of the estimator for the sample median proposed by varameter.
}
\details{
In this case we assume a Cauchy distribution and have an observed median, \eqn{m}, spread (interquartile range or range), \eqn{iqr}, and sample size, \eqn{n}.

We have the quantile function given by \eqn{Q(p) = \eta + \gamma \tan(\pi(p - 1/2))}, and the median is equal to the location parameter \eqn{\eta}. To estimate \eqn{\gamma}, we assume \eqn{iqr = Q(3/4) - Q(1/4)}, and solve to show \eqn{\gamma = iqr / 2}.

Then, to estimate the error of the sample median, we calculate \eqn{(2 \sqrt{n} g(m; \eta = m, \gamma = iqr / 2) )^{-1}}.
}
\seealso{
Other g_* Possible choices of assumed true density: exponential, Pareto,
Cauchy, and log-normal.: 
\code{\link{g_exp}()},
\code{\link{g_lnorm}()},
\code{\link{g_norm}()}
}
\concept{g_* Possible choices of assumed true density: exponential, Pareto,
Cauchy, and log-normal.}
