% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_exp.R
\name{g_exp}
\alias{g_exp}
\title{Estimator using the exponential distribution.}
\usage{
g_exp(n, median)
}
\arguments{
\item{n}{sample size}

\item{median}{median}
}
\description{
This estimator is for comparison purposes with the estimator proposed.
}
\details{
This estimator assumes we have observed a median, \eqn{m}, and sample size \eqn{n}. We naively assume an exponential distribution \eqn{\sim exp(\lambda)}, so that the median \eqn{\nu := log(2)/\lambda}.

This function then estimates \eqn{\lambda  :=  log(2) / m} and approximates the standard error of the sample median \eqn{(2 \sqrt{n} g(m; \lambda = log(2) / m))^{-1}} where \eqn{g} is the exponential density.
}
\seealso{
Other g_* Possible choices of assumed true density: exponential, Pareto,
Cauchy, and log-normal.: 
\code{\link{g_cauchy}()},
\code{\link{g_lnorm}()},
\code{\link{g_norm}()}
}
\concept{g_* Possible choices of assumed true density: exponential, Pareto,
Cauchy, and log-normal.}
