% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_lnorm.R
\name{g_lnorm}
\alias{g_lnorm}
\title{Estimator, assuming a log-normal distribution.}
\usage{
g_lnorm(n, median, spread, spread_type = "iqr")
}
\arguments{
\item{n}{sample size}

\item{median}{sample median}

\item{spread}{iqr or range value}

\item{spread_type}{iqr or range, defaults to iqr}
}
\description{
This estimator is for comparison purposes in the computational exploration of the estimator for the sample median proposed by varameter. In this case we assume a log-normal distribution
\eqn{ \sigma := \frac{1}{\Phi} }
}
\seealso{
Other g_* Possible choices of assumed true density: exponential, Pareto,
Cauchy, and log-normal.: 
\code{\link{g_cauchy}()},
\code{\link{g_exp}()},
\code{\link{g_norm}()}
}
\concept{g_* Possible choices of assumed true density: exponential, Pareto,
Cauchy, and log-normal.}
