% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g-norm.R
\name{g_norm}
\alias{g_norm}
\title{Estimator, assuming a normal distribution.}
\usage{
g_norm(n, median, spread, spread_type = "iqr")
}
\arguments{
\item{n}{sample size}

\item{median}{sample median}

\item{spread}{iqr or range value}

\item{spread_type}{iqr or range, defaults to iqr}
}
\description{
This estimator is for comparison purposes in the computational exploration of the estimator for the sample median proposed by varameter.
}
\details{
In this case we assume a normal distribution... todo complete this sometime soon.
}
\seealso{
Other g_* Possible choices of assumed true density: exponential, Pareto,
Cauchy, and log-normal.: 
\code{\link{g_cauchy}()},
\code{\link{g_exp}()},
\code{\link{g_lnorm}()}
}
\concept{g_* Possible choices of assumed true density: exponential, Pareto,
Cauchy, and log-normal.}
