% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hozo.R
\name{hozo}
\alias{hozo}
\alias{hozo_mean}
\alias{hozo_se}
\title{Estimate the Sample Mean from the Sample Median, Range, and Size}
\usage{
hozo_mean(a, m, b)

hozo_se(a, m, b, n)
}
\arguments{
\item{a}{Minimum value of sample.}

\item{m}{Median of sample.}

\item{b}{Maximum value of sample.}

\item{n}{Sample size.}
}
\description{
Estimate the Sample Mean from the Sample Median, Range, and Size
}
\seealso{
Other one_neet Inputs and outputs neet tested.: 
\code{\link{bland}},
\code{\link{effect_se}()},
\code{\link{g_}},
\code{\link{wan_c1}},
\code{\link{wan_c2}}

Other meansd_estimator Approximate an estimator for the mean from quartiles.: 
\code{\link{bland}},
\code{\link{wan_c1}},
\code{\link{wan_c2}}
}
\concept{meansd_estimator Approximate an estimator for the mean from quartiles.}
\concept{one_neet Inputs and outputs neet tested.}
