% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hozo_mean.R
\name{hozo_mean}
\alias{hozo_mean}
\title{Estimate the Sample Mean from the Sample Median, Range, and Size}
\usage{
hozo_mean(a, m, b)
}
\arguments{
\item{a}{Minimum value of sample.}

\item{m}{Median of sample.}

\item{b}{Maximum value of sample.}
}
\description{
See \code{\link{estimator_eqn_ref}} for details and references for equation
(3).
}
\seealso{
Other mean_estimator: 
\code{\link{wan_c3}}
}
\concept{mean_estimator}
