% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hozo_se.R
\name{hozo_se}
\alias{hozo_se}
\title{Estimate the Stardard Error of the Sample Mean from the Sample Median, Range,
and Size}
\usage{
hozo_se(a, m, b, n)
}
\arguments{
\item{a}{Minimum value of sample.}

\item{m}{Median of sample.}

\item{b}{Maximum value of sample.}

\item{n}{Sample size.}
}
\description{
See \code{\link{estimator_eqn_ref}} for details and references for equation
(6).
}
\seealso{
Other mean_error: 
\code{\link{bland_se}()},
\code{\link{wan_se_C1}()},
\code{\link{wan_se_C2}()},
\code{\link{wan_se_C3}()}
}
\concept{mean_error}
