% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdists.R
\docType{data}
\name{rdists}
\alias{rdists}
\title{List of distributions.}
\format{A list of densities and parameters.
\describe{
  \item{rdist}{A character indicating an `R`-readable density, such as `lnorm`.}
  \item{rpars}{A list containing parameters specific to that density.}
}}
\usage{
rdists
}
\description{
A list of densities and parameters used in the simulations.
}
\keyword{datasets}
