% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_par.R
\docType{data}
\name{sim_par}
\alias{sim_par}
\title{Things to investigate in the simulated data.}
\format{A list of variables.
\describe{
  \item{dist}{Distribution string in `R`-interpretable format. To convert the
  string back and forth from true name and `R`-interpretable, use the
  `dontpanic::dist_name` function.}
  \item{n}{Sample size for this set of trials.}
  \item{arm_adj}{Adjustment between first and second arm.}
  \item{trial}{Index of trial for this set of simulation parameters.}
}}
\usage{
sim_par
}
\description{
A list containing the various things I wanted to simulate over.
}
\keyword{datasets}
