% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varameta_package.R
\docType{package}
\name{varameta}
\alias{varameta}
\title{Estimators for Meta-Analysing Medians}
\description{
This package bundles together extension functions for the `R` package
`metafor`. There are estimators for meta-analysising medians, as well as
}
\details{
\tabular{ll}{ Package: \tab varameta\cr Type: \tab Package\cr License: \tab
GPL3\cr LazyLoad: \tab yes\cr }
}
\keyword{package}
