% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wan_c1.R
\name{wan_c1}
\alias{wan_c1}
\alias{wan_mean_C1}
\alias{wan_se_C1}
\title{Estimate the Sample Mean from Sample Median and Sample Range}
\usage{
wan_mean_C1(a, m, b)

wan_se_C1(a, b, n)
}
\arguments{
\item{a}{Minimum value of sample.}

\item{m}{Median of sample.}

\item{b}{Maximum value of sample.}

\item{n}{Sample size.}
}
\description{
Estimate the Sample Mean from Sample Median and Sample Range
}
\seealso{
Other one_neet Inputs and outputs neet tested.: 
\code{\link{bland}},
\code{\link{effect_se}()},
\code{\link{g_}},
\code{\link{hozo}},
\code{\link{wan_c2}}

Other meansd_estimator Approximate an estimator for the mean from quartiles.: 
\code{\link{bland}},
\code{\link{hozo}},
\code{\link{wan_c2}}
}
\concept{meansd_estimator Approximate an estimator for the mean from quartiles.}
\concept{one_neet Inputs and outputs neet tested.}
