% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wan_c2.R
\name{wan_c2}
\alias{wan_c2}
\alias{wan_mean_C2}
\alias{wan_se_C2}
\title{Estimate the Sample Mean from Quartiles, Range, and Size}
\usage{
wan_mean_C2(a, q_1, m, q_3, b)

wan_se_C2(a, q_1, m, q_3, b, n)
}
\arguments{
\item{a}{Minimum value of sample.}

\item{q_1}{First quartile of sample.}

\item{m}{Median of sample.}

\item{q_3}{Third quartile of sample.}

\item{b}{Maximum value of sample.}

\item{n}{Sample size.}
}
\description{
Estimate the Sample Mean from Quartiles, Range, and Size
}
\seealso{
Other one_neet Inputs and outputs neet tested.: 
\code{\link{bland}},
\code{\link{g_}},
\code{\link{hozo}},
\code{\link{wan_c1}}

Other meansd_estimator Approximate an estimator for the mean from quartiles.: 
\code{\link{bland}},
\code{\link{hozo}},
\code{\link{wan_c1}}
}
\concept{meansd_estimator Approximate an estimator for the mean from quartiles.}
\concept{one_neet Inputs and outputs neet tested.}
