% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wan_c3.R
\name{wan_c3}
\alias{wan_c3}
\alias{wan_mean_C3}
\alias{wan_se_C3}
\title{Estimate the Sample Mean from the Sample Median and Range}
\usage{
wan_mean_C3(q_1, m, q_3)

wan_se_C3(q_1, m, q_3, n)
}
\arguments{
\item{q_1}{First quartile of sample.}

\item{m}{Median of sample.}

\item{q_3}{Third quartile of sample.}

\item{n}{Sample size.}
}
\description{
Estimate the Sample Mean from the Sample Median and Range
}
\concept{mean_estimator}
