% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wan_mean_C1.R
\name{wan_mean_C1}
\alias{wan_mean_C1}
\title{Estimate the Sample Mean from Sample Median and Sample Range}
\usage{
wan_mean_C1(a, m, b)
}
\arguments{
\item{a}{Minimum value of sample.}

\item{m}{Median of sample.}

\item{b}{Maximum value of sample.}
}
\description{
See \code{\link{estimator_eqn_ref}} for details and references for equation
(3).
}
