% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wan_mean_C2.R
\name{wan_mean_C2}
\alias{wan_mean_C2}
\title{Estimate the Sample Mean from Quartiles, Range, and Size}
\usage{
wan_mean_C2(a, q_1, m, q_3, b)
}
\arguments{
\item{a}{Minimum value of sample.}

\item{q_1}{First quartile of sample.}

\item{m}{Median of sample.}

\item{q_3}{Third quartile of sample.}

\item{b}{Maximum value of sample.}
}
\description{
See \code{\link{estimator_eqn_ref}} for details and references for equation
(10).
}
\seealso{
Other mean_estimator: 
\code{\link{hozo_mean}()},
\code{\link{wan_mean_C1}()},
\code{\link{wan_mean_C3}()}
}
\concept{mean_estimator}
