% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wan_se_C1.R
\name{wan_se_C1}
\alias{wan_se_C1}
\title{Estimate the Standard Error of the Sample Mean from Sample Quartiles and Range}
\usage{
wan_se_C1(a, b, n)
}
\arguments{
\item{a}{Minimum value of sample.}

\item{b}{Maximum value of sample.}

\item{n}{Sample size.}
}
\description{
See \code{\link{estimator_eqn_ref}} for details and references for equation
(9).
}
