% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wan-se-C2.R
\name{wan_se_C2}
\alias{wan_se_C2}
\title{Estimate the Standard Error of the Sample Mean from the Sample Median, Range, and Size}
\usage{
wan_se_C2(a, q_1, m, q_3, b, n)
}
\arguments{
\item{a}{Minimum value of sample.}

\item{q_1}{First quartile of sample.}

\item{m}{Median of sample.}

\item{q_3}{Third quartile of sample.}

\item{b}{Maximum value of sample.}

\item{n}{Sample size.}
}
\description{
See \code{\link{estimator_eqn_ref}} for details and references for equation
(13).
}
\seealso{
Other mean_error: 
\code{\link{bland_se}()},
\code{\link{hozo_se}()},
\code{\link{wan_se_C1}()},
\code{\link{wan_se_C3}()}
}
\concept{mean_error}
