% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wan_se_C3.R
\name{wan_se_C3}
\alias{wan_se_C3}
\title{Estimate the Stardard Error of the Sample Mean from Sample Quartiles, Range, and Size}
\usage{
wan_se_C3(q_1, m, q_3, n)
}
\arguments{
\item{q_1}{First quartile of sample.}

\item{m}{Median of sample.}

\item{q_3}{Third quartile of sample.}

\item{n}{Sample size.}
}
\description{
See \code{\link{estimator_eqn_ref}} for details and references for equation
(16).
}
\seealso{
Other mean_error: 
\code{\link{hozo_se}()},
\code{\link{wan_se_C1}()},
\code{\link{wan_se_C2}()}
}
\concept{mean_error}
