% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bland.R
\name{bland}
\alias{bland}
\alias{bland_mean}
\alias{bland_se}
\title{Estimate Mean from Quartiles and Range (Bland)}
\usage{
bland_mean(a, q_1, m, q_3, b)

bland_se(a, q_1, m, q_3, b, n)
}
\arguments{
\item{a}{Minimum value of sample.}

\item{q_1}{First quartile of sample.}

\item{m}{Median of sample.}

\item{q_3}{Third quartile of sample.}

\item{b}{Maximum value of sample.}

\item{n}{Sample size.}
}
\description{
Method provided by Bland [todo: citation].
}
\seealso{
Other one_neet Inputs and outputs neet tested.: 
\code{\link{g_}},
\code{\link{hozo}},
\code{\link{wan_c1}},
\code{\link{wan_c2}}

Other meansd_estimator Approximate an estimator for the mean from quartiles.: 
\code{\link{hozo}},
\code{\link{wan_c1}},
\code{\link{wan_c2}}
}
\concept{meansd_estimator Approximate an estimator for the mean from quartiles.}
\concept{one_neet Inputs and outputs neet tested.}
