% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wan_mean_C3.R
\name{wan_mean_C3}
\alias{wan_mean_C3}
\title{Estimate the Sample Mean from the Sample Median and Range}
\usage{
wan_mean_C3(q_1, m, q_3)
}
\arguments{
\item{q_1}{First quartile of sample.}

\item{m}{Median of sample.}

\item{q_3}{Third quartile of sample.}
}
\description{
See \code{\link{estimator_eqn_ref}} for details and references for equation
(14).
}
\seealso{
Other mean_estimator: 
\code{\link{bland_mean}()},
\code{\link{hozo_mean}()},
\code{\link{wan_mean_C1}()},
\code{\link{wan_mean_C2}()}
}
\concept{mean_estimator}
